requirements.txt
vanna[gemini]>=2.0.0
fastapi
uvicorn[standard]
plotly
pandas
python-dotenv
google-genai
